update installer. create "beef" service account.
As part of the Linux install branch of the installer a systemd service is created. This commit adds the creation of a beef user account which is locked with no shell for use as a service account specific to beef. The service which is created by the installer has been updated to use the "beef" user.